refactor: Refactored Layers Architecture To Be A Spring Boot Application. Issue #2450 (#2469)
* #2450 1. Refactored Layers Architecture to be a Spring boot app. 2. Changed the xml-based configuration to annotation-based. 3. Used spring's constructor injection to pass around needed objects. 4. Implemented deleteAll() methods for the CakeBakingServiceImpl to be used during testing  5. Implemented a CommandLineRunner to run the spring boot app. And others.

* #2450 added the contents of the etc directory and a README.md

* #2450 made corrections in response to the PR tests

* #2450 made corrections in response to requested changes

* #2450 made corrections in response to requested changes

---------
