+++ layers/README.md
@@ -3,34 +3,31 @@ title: Layers
 category: Architectural
 language: en
 tag:
- - Decoupling
+- Decoupling
 ---
 
 ## Intent
 
-Layers is an architectural pattern where software responsibilities are divided among the different 
+Layers is an architectural pattern where software responsibilities are divided among the different
 layers of the application.
 
 ## Explanation
 
 Real world example
 
-> Consider a web site displaying decorated cakes for weddings and such. Instead of the web page 
-> directly reaching into the database, it relies on a service to deliver this information. The 
+> Consider a website displaying decorated cakes for weddings and such. Instead of the web page
+> directly reaching into the database, it relies on a service to deliver this information. The
 > service then queries the data layer to assimilate the needed information.
-
 In plain words
 
-> With Layers architectural pattern different concerns reside on separate layers. View layer is 
-> interested only in rendering, service layer assembles the requested data from various sources, and 
+> With Layers architectural pattern different concerns reside on separate layers. View layer is
+> interested only in rendering, service layer assembles the requested data from various sources, and
 > data layer gets the bits from the data storage.
-
 Wikipedia says
 
-> In software engineering, multitier architecture (often referred to as n-tier architecture) or 
-> multilayered architecture is a client–server architecture in which presentation, application 
+> In software engineering, multitier architecture (often referred to as n-tier architecture) or
+> multilayered architecture is a client–server architecture in which presentation, application
 > processing, and data management functions are physically separated.
-
 **Programmatic Example**
 
 On the data layer, we keep our cake building blocks. `Cake` consist of layers and topping.
@@ -38,14 +35,11 @@ On the data layer, we keep our cake building blocks. `Cake` consist of layers an
 ```java
 @Entity
 public class Cake {
-
   @Id
   @GeneratedValue
   private Long id;
-
   @OneToOne(cascade = CascadeType.REMOVE)
   private CakeTopping topping;
-
   @OneToMany(cascade = CascadeType.REMOVE, fetch = FetchType.EAGER)
   private Set<CakeLayer> layers;
 }
@@ -55,17 +49,11 @@ The service layer offers `CakeBakingService` for easy access to different aspect
 
 ```java
 public interface CakeBakingService {
-
   void bakeNewCake(CakeInfo cakeInfo) throws CakeBakingException;
-
   List<CakeInfo> getAllCakes();
-
   void saveNewTopping(CakeToppingInfo toppingInfo);
-
   List<CakeToppingInfo> getAvailableToppings();
-
   void saveNewLayer(CakeLayerInfo layerInfo);
-
   List<CakeLayerInfo> getAvailableLayers();
 }
 ```
@@ -74,20 +62,14 @@ On the top we have our `View` responsible of rendering the cakes.
 
 ```java
 public interface View {
-
   void render();
-
 }
-
 @Slf4j
 public class CakeViewImpl implements View {
-
   private final CakeBakingService cakeBakingService;
-
   public CakeViewImpl(CakeBakingService cakeBakingService) {
     this.cakeBakingService = cakeBakingService;
   }
-
   public void render() {
     cakeBakingService.getAllCakes().forEach(cake -> LOGGER.info(cake.toString()));
   }
