+++ layers/src/main/java/dao/CakeToppingDao.java
@@ -22,19 +22,18 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.exception;
+package dao;
+
+
+
+import entity.CakeTopping;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 /**
- * Custom exception used in cake baking.
+ * CRUD repository cake toppings.
  */
-public class CakeBakingException extends Exception {
+@Repository
+public interface CakeToppingDao extends JpaRepository<CakeTopping, Long> {
 
-  private static final long serialVersionUID = 1L;
-
-  public CakeBakingException() {
-  }
-
-  public CakeBakingException(String message) {
-    super(message);
-  }
 }
