+++ layers/src/main/java/dto/CakeInfo.java
@@ -22,68 +22,50 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.entity;
+package dto;
 
-import java.util.HashSet;
-import java.util.Set;
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
+import java.util.List;
+import java.util.Optional;
 
 /**
- * Cake entity.
+ * DTO for cakes.
  */
-@Entity
-public class Cake {
+public class CakeInfo {
 
-  @Id
-  @GeneratedValue
-  private Long id;
+    public final Optional<Long> id;
+    public final CakeToppingInfo cakeToppingInfo;
+    public final List<CakeLayerInfo> cakeLayerInfos;
 
-  @OneToOne(cascade = CascadeType.REMOVE)
-  private CakeTopping topping;
+    /**
+     * Constructor.
+     */
+    public CakeInfo(Long id, CakeToppingInfo cakeToppingInfo, List<CakeLayerInfo> cakeLayerInfos) {
+        this.id = Optional.of(id);
+        this.cakeToppingInfo = cakeToppingInfo;
+        this.cakeLayerInfos = cakeLayerInfos;
+    }
 
-  @OneToMany(cascade = CascadeType.REMOVE, fetch = FetchType.EAGER)
-  private Set<CakeLayer> layers;
+    /**
+     * Constructor.
+     */
+    public CakeInfo(CakeToppingInfo cakeToppingInfo, List<CakeLayerInfo> cakeLayerInfos) {
+        this.id = Optional.empty();
+        this.cakeToppingInfo = cakeToppingInfo;
+        this.cakeLayerInfos = cakeLayerInfos;
+    }
 
-  public Cake() {
-    setLayers(new HashSet<>());
-  }
+    /**
+     * Calculate calories.
+     */
+    public int calculateTotalCalories() {
+        var total = cakeToppingInfo != null ? cakeToppingInfo.calories : 0;
+        total += cakeLayerInfos.stream().mapToInt(c -> c.calories).sum();
+        return total;
+    }
 
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public CakeTopping getTopping() {
-    return topping;
-  }
-
-  public void setTopping(CakeTopping topping) {
-    this.topping = topping;
-  }
-
-  public Set<CakeLayer> getLayers() {
-    return layers;
-  }
-
-  public void setLayers(Set<CakeLayer> layers) {
-    this.layers = layers;
-  }
-
-  public void addLayer(CakeLayer layer) {
-    this.layers.add(layer);
-  }
-
-  @Override
-  public String toString() {
-    return String.format("id=%s topping=%s layers=%s", id, topping, layers.toString());
-  }
+    @Override
+    public String toString() {
+        return String.format("CakeInfo id=%d topping=%s layers=%s totalCalories=%d", id.orElse(-1L),
+                cakeToppingInfo, cakeLayerInfos, calculateTotalCalories());
+    }
 }
