+++ layers/src/main/java/dto/CakeLayerInfo.java
@@ -22,40 +22,39 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.dto;
+package dto;
 
 import java.util.Optional;
 
 /**
- * DTO for cake toppings.
+ * DTO for cake layers.
  */
-public class CakeToppingInfo {
+public class CakeLayerInfo {
 
-  public final Optional<Long> id;
-  public final String name;
-  public final int calories;
+    public final Optional<Long> id;
+    public final String name;
+    public final int calories;
 
-  /**
-   * Constructor.
-   */
-  public CakeToppingInfo(Long id, String name, int calories) {
-    this.id = Optional.of(id);
-    this.name = name;
-    this.calories = calories;
-  }
+    /**
+     * Constructor.
+     */
+    public CakeLayerInfo(Long id, String name, int calories) {
+        this.id = Optional.of(id);
+        this.name = name;
+        this.calories = calories;
+    }
 
-  /**
-   * Constructor.
-   */
-  public CakeToppingInfo(String name, int calories) {
-    this.id = Optional.empty();
-    this.name = name;
-    this.calories = calories;
-  }
+    /**
+     * Constructor.
+     */
+    public CakeLayerInfo(String name, int calories) {
+        this.id = Optional.empty();
+        this.name = name;
+        this.calories = calories;
+    }
 
-  @Override
-  public String toString() {
-    return String.format("CakeToppingInfo id=%d name=%s calories=%d",
-        id.orElse(-1L), name, calories);
-  }
+    @Override
+    public String toString() {
+        return String.format("CakeLayerInfo id=%d name=%s calories=%d", id.orElse(-1L), name, calories);
+    }
 }
