+++ layers/src/main/java/dto/CakeToppingInfo.java
@@ -22,16 +22,41 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.dao;
+package dto;
 
-import com.iluwatar.layers.entity.CakeLayer;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
 
 /**
- * CRUD repository for cake layers.
+ * DTO for cake toppings.
  */
-@Repository
-public interface CakeLayerDao extends CrudRepository<CakeLayer, Long> {
+public class CakeToppingInfo {
 
+    public final Optional<Long> id;
+    public final String name;
+    public final int calories;
+
+    /**
+     * Constructor.
+     */
+    public CakeToppingInfo(Long id, String name, int calories) {
+        this.id = Optional.of(id);
+        this.name = name;
+        this.calories = calories;
+    }
+
+    /**
+     * Constructor.
+     */
+    public CakeToppingInfo(String name, int calories) {
+        this.id = Optional.empty();
+        this.name = name;
+        this.calories = calories;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("CakeToppingInfo id=%d name=%s calories=%d",
+                id.orElse(-1L), name, calories);
+    }
 }
