+++ layers/src/main/java/exception/CakeBakingException.java
@@ -22,73 +22,22 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.entity;
+package exception;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
+import org.springframework.stereotype.Component;
 
 /**
- * CakeLayer entity.
+ * Custom exception used in cake baking.
  */
-@Entity
-public class CakeLayer {
+@Component
+public class CakeBakingException extends Exception {
 
-  @Id
-  @GeneratedValue
-  private Long id;
+    private static final long serialVersionUID = 1L;
 
-  private String name;
+    public CakeBakingException() {
+    }
 
-  private int calories;
-
-  @ManyToOne(cascade = CascadeType.ALL)
-  private Cake cake;
-
-  public CakeLayer() {
-  }
-
-  public CakeLayer(String name, int calories) {
-    this.setName(name);
-    this.setCalories(calories);
-  }
-
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public int getCalories() {
-    return calories;
-  }
-
-  public void setCalories(int calories) {
-    this.calories = calories;
-  }
-
-  @Override
-  public String toString() {
-    return String.format("id=%s name=%s calories=%d", id, name, calories);
-  }
-
-  public Cake getCake() {
-    return cake;
-  }
-
-  public void setCake(Cake cake) {
-    this.cake = cake;
-  }
+    public CakeBakingException(String message) {
+        super(message);
+    }
 }
