+++ layers/src/main/java/view/View.java
@@ -22,50 +22,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.dto;
-
-import java.util.List;
-import java.util.Optional;
+package view;
 
 /**
- * DTO for cakes.
+ * View interface.
  */
-public class CakeInfo {
+public interface View {
 
-  public final Optional<Long> id;
-  public final CakeToppingInfo cakeToppingInfo;
-  public final List<CakeLayerInfo> cakeLayerInfos;
+    void render();
 
-  /**
-   * Constructor.
-   */
-  public CakeInfo(Long id, CakeToppingInfo cakeToppingInfo, List<CakeLayerInfo> cakeLayerInfos) {
-    this.id = Optional.of(id);
-    this.cakeToppingInfo = cakeToppingInfo;
-    this.cakeLayerInfos = cakeLayerInfos;
-  }
-
-  /**
-   * Constructor.
-   */
-  public CakeInfo(CakeToppingInfo cakeToppingInfo, List<CakeLayerInfo> cakeLayerInfos) {
-    this.id = Optional.empty();
-    this.cakeToppingInfo = cakeToppingInfo;
-    this.cakeLayerInfos = cakeLayerInfos;
-  }
-
-  /**
-   * Calculate calories.
-   */
-  public int calculateTotalCalories() {
-    var total = cakeToppingInfo != null ? cakeToppingInfo.calories : 0;
-    total += cakeLayerInfos.stream().mapToInt(c -> c.calories).sum();
-    return total;
-  }
-
-  @Override
-  public String toString() {
-    return String.format("CakeInfo id=%d topping=%s layers=%s totalCalories=%d", id.orElse(-1L),
-        cakeToppingInfo, cakeLayerInfos, calculateTotalCalories());
-  }
 }
