+++ layers/src/test/java/com/iluwatar/layers/entity/CakeTest.java
@@ -31,6 +31,10 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.HashSet;
 import java.util.Set;
+
+import entity.Cake;
+import entity.CakeLayer;
+import entity.CakeTopping;
 import org.junit.jupiter.api.Test;
 
 /**
