+++ layers/src/test/java/com/iluwatar/layers/service/CakeBakingServiceImplTest.java
@@ -30,33 +30,56 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.iluwatar.layers.dto.CakeInfo;
-import com.iluwatar.layers.dto.CakeLayerInfo;
-import com.iluwatar.layers.dto.CakeToppingInfo;
-import com.iluwatar.layers.exception.CakeBakingException;
+import com.iluwatar.layers.app.LayersApp;
+import dto.CakeInfo;
+import dto.CakeLayerInfo;
+import dto.CakeToppingInfo;
+import exception.CakeBakingException;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import service.CakeBakingServiceImpl;
 import java.util.Collections;
 import java.util.List;
+
 import org.junit.jupiter.api.Test;
+
 
 /**
  * Date: 12/15/15 - 9:55 PM
  *
  * @author Jeroen Meulemeester
  */
+
+@SpringBootTest(classes = LayersApp.class)
 class CakeBakingServiceImplTest {
 
-  @Test
-  void testLayers() {
-    final var service = new CakeBakingServiceImpl();
+  private final CakeBakingServiceImpl cakeBakingService;
 
-    final var initialLayers = service.getAvailableLayers();
+  @Autowired
+  CakeBakingServiceImplTest(CakeBakingServiceImpl cakeBakingService) {
+    this.cakeBakingService = cakeBakingService;
+  }
+
+  @BeforeEach
+    void setUp() {
+        cakeBakingService.deleteAllCakes();
+        cakeBakingService.deleteAllLayers();
+        cakeBakingService.deleteAllToppings();
+    }
+
+
+  @Test
+
+  void testLayers() {
+    final var initialLayers = cakeBakingService.getAvailableLayers();
     assertNotNull(initialLayers);
     assertTrue(initialLayers.isEmpty());
 
-    service.saveNewLayer(new CakeLayerInfo("Layer1", 1000));
-    service.saveNewLayer(new CakeLayerInfo("Layer2", 2000));
+    cakeBakingService.saveNewLayer(new CakeLayerInfo("Layer1", 1000));
+    cakeBakingService.saveNewLayer(new CakeLayerInfo("Layer2", 2000));
 
-    final var availableLayers = service.getAvailableLayers();
+    final var availableLayers = cakeBakingService.getAvailableLayers();
     assertNotNull(availableLayers);
     assertEquals(2, availableLayers.size());
     for (final var layer : availableLayers) {
@@ -70,16 +93,14 @@ class CakeBakingServiceImplTest {
 
   @Test
   void testToppings() {
-    final var service = new CakeBakingServiceImpl();
-
-    final var initialToppings = service.getAvailableToppings();
+    final var initialToppings = cakeBakingService.getAvailableToppings();
     assertNotNull(initialToppings);
     assertTrue(initialToppings.isEmpty());
 
-    service.saveNewTopping(new CakeToppingInfo("Topping1", 1000));
-    service.saveNewTopping(new CakeToppingInfo("Topping2", 2000));
+    cakeBakingService.saveNewTopping(new CakeToppingInfo("Topping1", 1000));
+    cakeBakingService.saveNewTopping(new CakeToppingInfo("Topping2", 2000));
 
-    final var availableToppings = service.getAvailableToppings();
+    final var availableToppings = cakeBakingService.getAvailableToppings();
     assertNotNull(availableToppings);
     assertEquals(2, availableToppings.size());
     for (final var topping : availableToppings) {
@@ -93,28 +114,27 @@ class CakeBakingServiceImplTest {
 
   @Test
   void testBakeCakes() throws CakeBakingException {
-    final var service = new CakeBakingServiceImpl();
 
-    final var initialCakes = service.getAllCakes();
+    final var initialCakes = cakeBakingService.getAllCakes();
     assertNotNull(initialCakes);
     assertTrue(initialCakes.isEmpty());
 
     final var topping1 = new CakeToppingInfo("Topping1", 1000);
     final var topping2 = new CakeToppingInfo("Topping2", 2000);
-    service.saveNewTopping(topping1);
-    service.saveNewTopping(topping2);
+    cakeBakingService.saveNewTopping(topping1);
+    cakeBakingService.saveNewTopping(topping2);
 
     final var layer1 = new CakeLayerInfo("Layer1", 1000);
     final var layer2 = new CakeLayerInfo("Layer2", 2000);
     final var layer3 = new CakeLayerInfo("Layer3", 2000);
-    service.saveNewLayer(layer1);
-    service.saveNewLayer(layer2);
-    service.saveNewLayer(layer3);
+    cakeBakingService.saveNewLayer(layer1);
+    cakeBakingService.saveNewLayer(layer2);
+    cakeBakingService.saveNewLayer(layer3);
 
-    service.bakeNewCake(new CakeInfo(topping1, List.of(layer1, layer2)));
-    service.bakeNewCake(new CakeInfo(topping2, Collections.singletonList(layer3)));
+    cakeBakingService.bakeNewCake(new CakeInfo(topping1, List.of(layer1, layer2)));
+    cakeBakingService.bakeNewCake(new CakeInfo(topping2, Collections.singletonList(layer3)));
 
-    final var allCakes = service.getAllCakes();
+    final var allCakes = cakeBakingService.getAllCakes();
     assertNotNull(allCakes);
     assertEquals(2, allCakes.size());
     for (final var cakeInfo : allCakes) {
@@ -130,61 +150,49 @@ class CakeBakingServiceImplTest {
 
   @Test
   void testBakeCakeMissingTopping() {
-    final var service = new CakeBakingServiceImpl();
-
     final var layer1 = new CakeLayerInfo("Layer1", 1000);
     final var layer2 = new CakeLayerInfo("Layer2", 2000);
-    service.saveNewLayer(layer1);
-    service.saveNewLayer(layer2);
+    cakeBakingService.saveNewLayer(layer1);
+    cakeBakingService.saveNewLayer(layer2);
 
     final var missingTopping = new CakeToppingInfo("Topping1", 1000);
-    assertThrows(CakeBakingException.class, () -> {
-      service.bakeNewCake(new CakeInfo(missingTopping, List.of(layer1, layer2)));
-    });
+    assertThrows(CakeBakingException.class, () -> cakeBakingService.bakeNewCake(new CakeInfo(missingTopping, List.of(layer1, layer2))));
   }
 
   @Test
   void testBakeCakeMissingLayer() {
-    final var service = new CakeBakingServiceImpl();
-
-    final var initialCakes = service.getAllCakes();
+    final var initialCakes = cakeBakingService.getAllCakes();
     assertNotNull(initialCakes);
     assertTrue(initialCakes.isEmpty());
 
     final var topping1 = new CakeToppingInfo("Topping1", 1000);
-    service.saveNewTopping(topping1);
+    cakeBakingService.saveNewTopping(topping1);
 
     final var layer1 = new CakeLayerInfo("Layer1", 1000);
-    service.saveNewLayer(layer1);
+    cakeBakingService.saveNewLayer(layer1);
 
     final var missingLayer = new CakeLayerInfo("Layer2", 2000);
-    assertThrows(CakeBakingException.class, () -> {
-      service.bakeNewCake(new CakeInfo(topping1, List.of(layer1, missingLayer)));
-    });
+    assertThrows(CakeBakingException.class, () -> cakeBakingService.bakeNewCake(new CakeInfo(topping1, List.of(layer1, missingLayer))));
   }
 
   @Test
   void testBakeCakesUsedLayer() throws CakeBakingException {
-    final var service = new CakeBakingServiceImpl();
-
-    final var initialCakes = service.getAllCakes();
+    final var initialCakes = cakeBakingService.getAllCakes();
     assertNotNull(initialCakes);
     assertTrue(initialCakes.isEmpty());
 
     final var topping1 = new CakeToppingInfo("Topping1", 1000);
     final var topping2 = new CakeToppingInfo("Topping2", 2000);
-    service.saveNewTopping(topping1);
-    service.saveNewTopping(topping2);
+    cakeBakingService.saveNewTopping(topping1);
+    cakeBakingService.saveNewTopping(topping2);
 
     final var layer1 = new CakeLayerInfo("Layer1", 1000);
     final var layer2 = new CakeLayerInfo("Layer2", 2000);
-    service.saveNewLayer(layer1);
-    service.saveNewLayer(layer2);
+    cakeBakingService.saveNewLayer(layer1);
+    cakeBakingService.saveNewLayer(layer2);
 
-    service.bakeNewCake(new CakeInfo(topping1, List.of(layer1, layer2)));
-    assertThrows(CakeBakingException.class, () -> {
-      service.bakeNewCake(new CakeInfo(topping2, Collections.singletonList(layer2)));
-    });
+    cakeBakingService.bakeNewCake(new CakeInfo(topping1, List.of(layer1, layer2)));
+    assertThrows(CakeBakingException.class, () -> cakeBakingService.bakeNewCake(new CakeInfo(topping2, Collections.singletonList(layer2))));
   }
 
 }
